requirements.txt
# Deep learning
torch>=2.0.0
torchvision>=0.15.0

# Computer vision
opencv-python>=4.8.0
pillow>=10.0.0

# Data & ML utilities
numpy>=1.24.0
pandas>=2.0.0
scikit-learn>=1.3.0

# Visualisation
matplotlib>=3.7.0
seaborn>=0.12.0
plotly>=5.15.0

# Web app
streamlit>=1.28.0
